refactor: feat/QuantLabPage 리팩토링
