+++ src/components/FormControlGroup.tsx
@@ -7,7 +7,7 @@ interface TmpFormProps {
 
 // NOTE: FormControl, FormGroup 등의 역할을 아직 모르겠다. 그냥 CheckBox 들을 배치하는거랑 무슨 차이?
 
-export default function TmpForm({ children }: TmpFormProps) {
+export default function FormControlGroup({ children }: TmpFormProps) {
   return (
     <FormControl sx={{ m: 3 }} component="fieldset" variant="standard">
       {/* <FormLabel component="legend">재무상태</FormLabel> */}
