--- src/components/ModalWithButton/index.ts
@@ -1,3 +1,0 @@
-import ModalWithButton from "./ModalWithButton";
-
-export default ModalWithButton;
