+++ src/pages/MainPage/Selecter/TermSelect.tsx
