+++ src/pages/QuantLabPage/QuantModelCreation/LabModal/LabModal.tsx
@@ -1,8 +1,8 @@
 import { Checkbox, FormControlLabel } from "@mui/material";
 import { useState } from "react";
 
-import TmpForm from "../../../../components/FormControlGroup/FormControlGroup";
-import ModalWithButton from "../../../../components/ModalWithButton/ModalWithButton";
+import FormControlGroup from "../../../../components/FormControlGroup";
+import ModalWithButton from "../../../../components/ModalWithButton";
 import { IBusinessArea } from "../QuantModelCreation";
 
 interface QuantLabModalProps {
@@ -43,7 +43,7 @@ export default function QuantLabModal({
         }
         label="Select all"
       />
-      <TmpForm>
+      <FormControlGroup>
         {Object.keys(state).map((key, idx) => {
           return (
             <FormControlLabel
@@ -59,7 +59,7 @@ export default function QuantLabModal({
             />
           );
         })}
-      </TmpForm>
+      </FormControlGroup>
     </ModalWithButton>
   );
 }
