+++ src/pages/QuantLabPage/QuantModelTable/QuantModelTable.tsx
@@ -10,7 +10,7 @@ import {
 } from "@mui/x-data-grid";
 import { useCallback, useMemo } from "react";
 
-import { KoreanName } from "../constants";
+import { VariableNameTranslate } from "../constants";
 import { IModel } from "../QuantLabPage";
 import QuantLabSaveModal from "../QuantModelCreation/LabModal/LabSaveModal";
 
@@ -28,10 +28,10 @@ export default function QuantModelTable({
   const columns: GridColDef[] = FIELDS.map((val) => {
     return {
       field: val,
-      headerName: KoreanName[`${val}` as keyof typeof KoreanName],
+      headerName:
+        VariableNameTranslate[`${val}` as keyof typeof VariableNameTranslate],
       width: 150,
     };
-    // return { field: val, headerName: val, width: 150 };
   });
 
   const deleteModel = useCallback(
@@ -93,7 +93,6 @@ export default function QuantModelTable({
             rows={rows}
             columns={columnsWithButton}
             onSelectionModelChange={(model) => {
-              // console.log("hi...", model);
               setSelectionModel(model);
             }}
           />
@@ -110,9 +109,4 @@ const FIELDS = [
   "winning_percentage",
   "max_loss_rate",
   "holdings_count",
-]; // NOTE: 적절한가? 잘 모르겠음
-/* TODO
- * 데이터 구조, 상수화 등에 대한 고민
- * 폭 개선 (디자인 개선)
- * 버튼 추가
- */
+];
