+++ src/pages/QuantLabPage/QuantModelViewer/QuantModelViewer.tsx
@@ -1,37 +1,22 @@
 import { Paper } from "@mui/material";
-import { GridSelectionModel } from "@mui/x-data-grid";
 import styled from "styled-components";
 
-import Graph from "../../../components/graph";
-import { IModel } from "../QuantLabPage";
+import { IChart } from "../QuantLabPage";
+import Chart from "./Chart";
 
-export interface tmpModel {
-  id: number;
-  model_name: string;
-  chart_data: {
-    start_date: string;
-    profit_kospi_data: number[];
-    profit_rate_data: number[];
-  };
-}
 interface ChartViewerProps {
-  models: tmpModel[];
+  charts: IChart[];
 }
 
-export default function ChartViewer({ models }: ChartViewerProps) {
+export default function ChartViewer({ charts }: ChartViewerProps) {
   return (
     <MainContainer variant="outlined">
-      {/* <Example /> */}
-      <Graph models={models} />
+      <Chart charts={charts} />
     </MainContainer>
   );
 }
 
 const MainContainer = styled(Paper)`
-  // border: 3px solid green;
   width: 79%;
-
   padding: 20px;
 `;
-
-// TODO: 아직 상태 적용 X
