+++ src/pages/QuantLabPage/constants.ts
@@ -1,5 +1,5 @@
 // export const enum KoreanName {
-export enum KoreanName {
+export enum VariableNameTranslate {
   "model_name" = "모델명",
   "cumulative_return" = "누적 수익률",
   "annual_average_return" = "연평균 수익",
